+++ lists/tests/test_views.py
@@ -1,5 +1,4 @@
 from django.core.urlresolvers import resolve
-from django.template.loader import render_to_string
 from django.test import TestCase
 from lists.views import home_page
 from django.http import HttpRequest
@@ -21,35 +20,6 @@ class HomePageTest(TestCase):
         #self.assertTrue(response.content.startswith(b'<html>'),msg='ACTUAL CONTENT: {}'.format(response.content))
         self.assertIn(b'<title>', response.content)
         #self.assertTrue(response.content.endswith(b'</html>'))
-
-
-class ListandItemModelsTest(TestCase):
-
-    def test_saving_and_retrieving_items(self):
-        list_ = List()
-        list_.save()
-
-        first_item = Item()
-        first_item.text = 'The first (ever) list item'
-        first_item.list = list_
-        first_item.save()
-
-        second_item = Item()
-        second_item.text = 'Item the second'
-        second_item.list = list_
-        second_item.save()
-
-        saved_items = Item.objects.all()
-        saved_list = List.objects.first()
-        self.assertEqual(saved_list, list_)
-        self.assertEqual(saved_items.count(), 2)
-
-        first_saved_item = saved_items[0]
-        second_saved_item = saved_items[1]
-        self.assertEqual(first_saved_item.text, 'The first (ever) list item')
-        self.assertEqual(first_saved_item.list, list_)
-        self.assertEqual(second_saved_item.text, 'Item the second')
-        self.assertEqual(second_saved_item.list, list_)
 
 
 class ListViewTest(TestCase):
